feat: Double support and round support.

+++ src/main/java/link/botwmcs/economic/capability/entity/IPlayerEcoData.java
@@ -3,8 +3,8 @@ package link.botwmcs.economic.capability.entity;
 import dev.onyxstudios.cca.api.v3.entity.PlayerComponent;
 
 public interface IPlayerEcoData extends PlayerComponent {
-    Integer getMoney();
+    Double getMoney();
     Boolean getIsFirstJoin();
-    void setMoney(Integer money);
+    void setMoney(Double money);
     void setIsFirstJoin(Boolean isFirstJoin);
 }
